+++ src/main/java/com/elmakers/mine/bukkit/plugins/magic/spells/FillSpell.java
@@ -7,8 +7,6 @@ import org.bukkit.block.Block;
 import com.elmakers.mine.bukkit.blocks.BlockList;
 import com.elmakers.mine.bukkit.blocks.FillBatch;
 import com.elmakers.mine.bukkit.blocks.MaterialBrush;
-import com.elmakers.mine.bukkit.effects.EffectUtils;
-import com.elmakers.mine.bukkit.effects.ParticleType;
 import com.elmakers.mine.bukkit.plugins.magic.BrushSpell;
 import com.elmakers.mine.bukkit.plugins.magic.SpellResult;
 import com.elmakers.mine.bukkit.plugins.magic.TargetType;
@@ -71,8 +69,7 @@ public class FillSpell extends BrushSpell
 			
 			FillBatch batch = new FillBatch(this, secondLocation, targetBlock.getLocation(), buildWith);
 
-			int maxDimension = parameters.getInteger("max_dimension", DEFAULT_MAX_DIMENSION);
-			
+			int maxDimension = parameters.getInteger("max_dimension", DEFAULT_MAX_DIMENSION);		
 			maxDimension = (int)(mage.getConstructionMultiplier() * maxDimension);
 			
 			if (!batch.checkDimension(maxDimension))